Rework public API
So far the public API had been comprised of a single function, run(),
and the crate had not actually been meant to be used as a library.
With this patch we change this fact. Specifically, we introduce proper
public function for capturing a screenshot, along with an "option" type
for configuring certain details of the process. This function is now
what is being used by the binary.