Properly isolate chromedriver instances
Our chromedriver instances are not exactly isolated with respect to each
other, causing all sorts of havoc once we attempt to take multiple
screenshots concurrently.
There are mainly two reasons for this:
1) we use a fixed hardcoded port, meaning that all invocations end up
   with the same chromedriver instance by design
2) all our chromedriver instances are started with the same session data
   directory; because the directory contains a shared socket, all
   instances end up working on shared data

This change fixes both problems. We start chromedriver with --port=0,
which causes it to use a kernel assigned port. Unfortunately, there is
not really a way to retrieve this port, other than wrangling with Linux
specific details. So we end up finding all the process' sockets bound to
local host and use the port of the first one.
In order to isolate session data, we use a new temporary directory for
each instance, and provide the path to it to each instance.