Factor out a proper crate for shavecli
With the recent focus on having a proper public API for the library
portion of the crate, it makes less and less sense to have the library
and the command line application combined in a single crate. That will
just cause unnecessary dependencies all over the place.
With this change we move the CLI functionality into a new workspace
member, shavecli.