Add support for removing elements before screenshot
Sometimes it is necessary to remove one or more elements before taking a
screenshot, e.g., because they may mess up the output. To make that
possible in a more or less generic fashion, this change allows passing
in a CSS selector of describing the elements to remove.

Note that WebDriver doesn't really have any APIs for modifying the DOM.
There is some "element clear" functionality, but it mostly applies to
resetting the state of input fields to "empty". To make it work
nevertheless, we just execute a short Javascript snippet in the context
of the window being screenshot.