Add support for specifying window size to use
Depending on which part of a website is being screenshot and how the
website is laid out to begin with, having a default window size may
prove very limiting as it may result in awkward looking screenshot
contents. It is unlikely that there is any automated approach for
achieving "best" results.
As such, this change introduces the means for users to simply specify
the window size to use.