Factor out dedicated Client type
The single-public-function that we had so far basically constitutes
the minimum viable API for providing our desired functionality. However,
it is not particularly flexible or performant. Consider for example the
case of a user wanting to capture multiple screenshots. Right now each
invocation spins up a new chromedriver process and instantiates a new
webdriver client speaking to it, which is expensive.
To mitigate these problems, this change moves the screenshot
functionality into the newly introduced Client type. An object of this
type basically encapsulates a chromedriver process alongside a webdriver
client, both of which can be reused for issuing multiple requests. On
top of that, in the future we can more easily add a "builder" style type
on top of that, which enables more configurability of this client.