Stop leaking Chrome data directory
In order to isolate Chrome instance somewhat, we start each with a new
data directory. However, the temporary directory we created so far was
dropped while the browser was still running and had files from it
opened. As a result, the temporary directory was never removed as it
should have been. Fix this problem