Stop passing --disable-dev-shm-usage to Chrome
Disabling the usage of /dev/shm is likely to result in additional disk
I/O, which may already be a bottleneck. It also apparently mostly helps
in Docker environment, which we couldn't care less about. Remove usage
of the flag as it seems to speed up general all things webdriver.