Disable some threaded functionality
Pass some more additional options to the Chrome instance being launched
to disable some threaded functionality in an attempt to consume fewer
resources. That being said, we have seen issues with the
--disable-threaded-compositing option, where shave functionality was
completely broken, with both rendering as well as CSS selector matching
not working. So refrain from using it, but add a note about that.