cli: Introduce 'launch' command
Introduce the 'launch' command to start up Chrome as automation would,
except not in headless mode but so that a user can actually interact
with it. This command is mostly meant for debugging matters, where it
can be of importance to start the browser with the exact options that
shave itself would.